Added `add_reeds_columns` utility and re-organized files
